Exclude env variables quotes when importing openrc
This fixes Issue #67.

Previously, the `Regex.regex` was doing a match/capture on everything
following the `=`, so the match includes `"around/value"`. This alters
the regex to only match what is contained within "double quotes".

There is no "./tests" directory, or I would include unit tests to verify
that the change is valid and adheres to expectations.

Refactor note: since `processOpenRc` does *not* return a Model nor a
(Model, Cmd Msg) tuple, the `model` argument could be narrowed to just
to accept just `model.creds` to use when leverage _previous_ value in
the records "oldField"s.

Can contribute a patch with this change if desired.