Ensure variations of OS_PROJECT_DOMAIN_* matched
It appears in the example that was provided by Julian there is a new
variable name for setting the name of a Project Domain:

    $OS_PROJECT_DOMAIN_ID

The regex for is defined to allow both `_ID` and previous `_NAME`.

This is achieved by use a non-capture group, defined with the `?:`
syntax. If you used a capture group, it would mean that `newField`
would populate `Creds` with either "_ID" or "_NAME" - totally not what
_anyone_ wants. So the resulting regular expression is:

"export OS_PROJECT_DOMAIN(?:_NAME|_ID)=\"?([^\"\n]*)\"?"

This is also included "not" matchin newline characters. I call out the
syntax here because I was not hip to ability for Elm (or JS) to define
a non-capture group (so much :claps: for Regex authors).