Handle new project domain variable suffix
The latest API Version 3 "openrc" has the project domain name stored in
the enviroment variable: `OS_PROJECT_DOMAIN_ID` [0]

This commit defines the regular expression with a non-capture group such
that the "matches" are not thrown off, or require special casing.

I did not know that Elm's Regex module had this ability.

The syntax is: (?:_NAME|_ID)

What makes it a non-capture is the preceding `?:` within the grouping
parens.

[0] https://gist.github.com/lenards/ad95edd90775a53d9d5b26236a968428#file-openrc-test-example-sh