Problem: No automated quality checks on MRs
Since moving to GitLab we don't have automatic testing. We used to have this on GitHub thanks to TravisCI.

Solution: Use GitLab to run the build and test steps using GitLab CI/CD on every MR.

Closes #114

This also closes #84 - because I'm using the build artifacts of the `elm_make` job (index.html) in the `pages` job.