Problem: Buttons look washed out.
Due to using PatternFly grayscale palette which is slightly offset from
the default elm-style-framework one.

We could try to use the PatternFly grayscale palette, and 'slide' it
down so that things are a bit darker.

Solution: For now just revert back to using the default grayscale
palette provided by elm-style-framework