Problem: BadPayload error when creating a new server
Due to the server creation endpoint only returning basic attributes,
not full server details.

Solution: Just decode a 'ServerUuid' instead of a full server

We're not using the server details in the 'receiveCreateServer' function
anyway - we're just requesting all the servers & networks and changing
the view to the server list.