adding dockerfile to build browser binary
It could be the case that the user(developer) wants a simple environment
to test or otherwise develop with exosphere, but does not trust node
or an app like electron. This provides a simple Dockerfile
to build the code into, along with instructions in the README for building
and interacting with the containers.

Signed-off-by: vsoch <vsochat@stanford.edu>