Consolidate parameters of ListProjectServers view into ServerListViewParams record (part 1 of 2)
In advance of moving where we store "selected-ness" of a Server (from that Server's exoProps to a list of Server UUIDs stored as a view parameter), this change will avoid a long (many-variabled) type definition for the ListProjectServers view.

This commit will not compile, it just changes the types. Next commit will fix the code to use the updated types.