Consolidate parameters of ListProjectServers view into ServerListViewParams record (part 2 of 2)
In advance of moving where we store "selected-ness" of a Server (from that Server's exoProps to a list of Server UUIDs stored as a view parameter), this change will avoid a long (many-variabled) type definition for the ListProjectServers view.

This commit fixes the code that was broken by the changes to types in the previous commit.