Define Quota/Usage View
Relates to Issue #359

Defining the beginning of an information view that shows the "in use"
and "limits" associated with the current project being viewed.

Here, the information is shown only in a textual rendering (no graphs,
etc). There is also no delineation between "my usage" and the overall
usage of the members of the _grouping_ (i.e. Project).

Surfacing the information, even in a crude form, appears to be helpful.