Match 960px width of Servers view
This switches from a wrappedRow and overflows (and creating an x-axis
scrollbar situation) to a fill, column that has a min-width of 960px.

This approach mean less informating rendered "above the fold", yet if
we couple this with a forth-coming change of "filter by text" match,
a la Issue #101's suggestion (re PatternFly example) then we can have
consistent layout placement, and quick access.