Ensure "description" does not overflow volume card
The rendering of a volume card is roughly 906 px when considering the
padding after applying a min-width of 960. The value part of the kv row
then shouldn't render wider than 706px, and we want the text to flow
like a paragraph, so we change to `Element.paragraph` over Element.text.