Fixed accordion direction for IP details
Should point right when not expanded, and down when expanded.

In the process I had to change to using an icon instead of a string,
and I decided to keep the text of the accordion as 'IP Details', instead
of switching back and forth between 'IP Details' and 'IP Summary'.