Make palette, logo, and app title configurable at runtime via flags
Deployer/user can define a primary and secondary color, custom logo, and application title. This allows for custom "white labeling".

The entire color palette is no longer stored in the model. Instead we just store the primary/secondary colors and style mode (light or dark), and derive the full color palette from that dynamically. This means that a Style.Types.ExoPalette is now generally passed around the view functions, instead of a Types.Types.Style.