Added 'additionalProperties' to Image
Decoding this dictionary of strings from all image properties which are
NOT present in a list of names of base properties. This list of base
property names is currently hard-coded and should be loaded from the
Glance image schema endpoint in the future.

This solution was based on a post on Elm Discourse:

https://discourse.elm-lang.org/t/how-to-write-a-generic-json-decoder/3179