Routing functions: Couple key and path prefix to app state
In this patch I'm proposing a small change to the routing functions
in preparation for work reorganizing the view context. Why? What's
the big deal?

In all existing calls to `pushUrl`, `replaceUrl`, and `withReplaceUrl`
we are extracting the same two `navigationKey` and `urlPathPrefix`
properties, passing them in, and doing the same logic. This gets
duplicated at every call site.

Given that it's clear this is a common-enough idiom in the app it
warrants an abstraction that ties "view-related information" with
the generated URLs, and I'm proposing we do that by passing the
view info directly.

In this patch we're using an extensible record type as the input
parameter. We could use something like `SharedModel` or a new type
alias only specifying the properties we want, but we cannot then
pass `SharedModel` in all cases without a type error. Eventually
I propose that we pass `ViewContext`, because if indeed the URL
routing is based on `ViewContext` then I don't think we should
arbitrarily constraint that we cannot base it off of _other_ view
parameters than the two we currently do. At that time, when the
`ViewContext` is also reorganized, we can type these functions
to accept `ViewContext` instead of `{ r | navigationKey... }`

There should be no visual or functional changes in this PR.