Style refactor: Favor single list iteration and pattern-matching
Previously we were imperatively constructing the last and middle
elements in the breadcrump by reversing the items twice and handling
an array of possible list lengths.

In this patch we're declaratively deconstructing the list such that
we never have to deal with an empty list tail, and such that we can
directly map the empty list vs. non-empty list cases.

There should be no visual or functional changes in this patch.