View context state: move into shared model state
Previously the view context was pulled out as a separate concept
on every render, being generated dynamically from `sharedModel`.
Surprisingly though it didn't depend on being calculated dynamically
and really was the product of the initial bootup _plus_ changes to
the style as configured through settings, a change that already trickles
through the main app update function.

In this patch we're moving `viewContext` and some associated values
out of the `sharedModel` and into a new property _of_ `sharedModel`
called `viewContext`. This we pass around the app vs. the dynamically
calculated value `context` from before.

This change makes it more straightforward to set view context properties
on the intial app boot and it removes a level of duplication present in
the app. This duplication was previously confusing because many functions
received `sharedModel` and also `context`, exposing the fact that it could
have gotten information like `urlPathPrefix` from either of those sources.

One question this raises is the role and use of `cloudSpecificConfigs`
which doesn't inherently sound like it's only a view concern, but has
moved into `viewContext` because that's where it's mostly used.