View: Format numbers according to inferred browser locale
In this patch we're formatting displayed numbers according to an
inferred locale at app boot. The inferred locale is somewhat goofy
because browsers natively support various number systems and display
contexts, but Elm doesn't expose the `Intl` API and so we can't
access it other than via a port.

The `FormatNumber` library therefore makes a first-order guess at the
configured locale when given an example string that's sufficiently
large and negative, using string matching to guess the digit separator,
groupings, and decimal separator. To that end we don't have full
numerical formatting support but a baseline is a good improvement here,
if for no other reason that we group thousands with a separator.

The API data showing instance sizes is not localized due to oddities
in the reported values when compared to use expectations. This patch
chooses not to format those numbers to avoid confusing common sizes,
such as "8 GB", with the more technically correct but unfamilar "9 GB"
it actually represents, or the "8 GiB" that isn't as common of a
reporting unit.