Add `Helpers.List.multiSortBy` for multi-level sorting
This helper function replaces multi-pass sorting in existing code
in favor of a single sort pass.

```elm
restaurants
    |> List.sortBy .rating
    |> List.sortBy .isOpen
```

This construction requires separate iteration and also requires
thinking backwards about the priority of comparisons in the sort.
Whatever is the primary sort value has to come last.

With this new function we can avoid multiple sort passes and
specify the priority of the sort values based on their order
in the supplied list, with the first being the primary.

```elm
restaurants
    |> Helpers.List.multiSortBy [.isOpen, .rating]
```