Refactor: Replace superfluous closures with inline code
There are places in the existing code where we wrap functions in
functions that essentially do nothing, but these cases create
needless functions and function calls in the built JS file.

While not particularly harmful there's no need to be extra verbose
and it's generally good practice to prefer a construction that doesn't
involve additional overhead if the alternatives are roughly equal
in terms of developer cost and code clarity.

In this patch we make the case that `map Tag value` is such a case
when compared against the pattern of `map (\a -> Tag a) value`

This patch shrinks the output build by 2.3 kB before minification/gzip.