Refactor: Skip double-iterating over string to replace values
Previously, the `novaUrlWithMicroversionSupport` helper function
gated a string replacement with a check if the string to be replaced
is found before actually replacing it. This leads to scanning the
string more than necessary since the search is implicit in the
replacement.

In this patch we're removing the checks and unconditionally running
the substitution. If the target string is missing, nothing will be
substituted and the code should be functionally equivalent to the
more verbose construction.

There appears to be more actual overhead in the Elm implementation
of string replacement than is necessary and so this may actually be
slower, though insignificantly so given how short the strings are.
I find the code clearer without the conditionals, you may not 🤷.