Refactor `newServerNetworkOptions` to flatten decision tree
In `newServerNetworkOptions` we build a deeply-nested decision tree
for reporting new server network options. While readble from top to
bottom, the nesting can make it difficult to physically read the code.

In this patch we're flattening the tree by coupling items in the `case`
structure and using `Maybe` functions in place of custom `case` matches.

There should be no functional or visual changes in this patch.