Refactor: Flatten logic in `serverNeedsFrequentPoll`
Previously the logic in `serverNeedsFrequentPoll` relied on a
deeply-nested decision tree. While physically difficult at times to
read with a reasonable-but-large font, this risked obscuring some
of what goes into the decision for whether or not a frequent poll
is required for a server.

In this patch we're flattening the logic through coupling multiple
values into a single `case` match and by pattern-matching not only
on the outer type tag, but also on specific values of the type.

There should be no functional or visual changes in this patch.