Fail on app startup when cloud config is invalid
The success criteria for issue 582 state that the app should "[die]
immediately and loudly" when flag decoding fails. As far as I can
tell, the only data which might need to be validated are Flags.clouds
and Flags.storedData, as these are the only members which are JSON
values. As all other types are primitive, errors here would be type
errors which the Elm runtime would handle. Flags.storedState uses a
default empty state when decoding fails, and does not log any errors,
so I assume that errors in this data should not "crash" the app.