Widgets: Create Home Logo widget
Previously the Nav page created the markup and styling for the home logo
inline with all the other logic of global navigation.

In this patch we're creating a new widget, the "home logo" that routes
to `Home` but allows for passing in a URL for the logo and an app title.

To support this componentization of the home logo we're also introducing
`View.Helpers.renderIf` which removes a bit of noise when conditionally
rendering a component based on some condition.

This patch is part of preparation work for #534 and exploratory for how
to start reorganizing the menu components for easier modularization and
reuse in new contexts (sidebar > context selector).