Widgets: Create `navButton` and `iconButton` to modularize menus
Prep work for #534

In this patch we're creating abstractions for the icon button and
also for the nav button. The icon button is used in several places
around the app and I think we can share that component with the
ones in the menu. The nav button is a small wrapper around the
icon button that adds nav-menu-specific behaviors, specifically
the routing.

The sidebar -> project selector will likely use these icon buttons
if the component API works fine, or they will modify the API when
it comes time to refactor them.