Updates `npm` scripts to smooth issues when packages aren't already available
Previously `elm-analyse` wasn't called out as a dependency and so if you
attempted to run it without manually installing it would fail with an error.

In this patch we're adding it as a `devDependency` so that it's available
automatically after running `npm install`.

To further cleanup this project we're calling `npm` packages in the scripts
through `npx` so that if they aren't already installed it won't fail. We're
also adding an alias `start` for consistency with other `npm` scripts and
local development flows.