Feature: Add logout button and functionality
Previously the only way to "logout" was to remove all projets, but this left
persistent state in the browser and in local storage and it wasn't clear that
this is how one is supposed to logout.

In this patch we're adding a new nav button to logout which when clicked will
clear all local app state, wipe it from the browser, and reload the page. This
adds a common idiom and should give people confidence that they can safely use
this app on a public computer.

The logout button is always visible, which is a bit strange, but that's because
the app doesn't particalary have a logged-out state. If at some point we build
that functionality then we can update the logout button to respond to that.