Refactor: Use list find/any/all functions
"Write what you want, not how to get it."

In this patch we're refactoring instances of `List.filter |> List.head`
to use `List.Extra.find` directly, which indicates purpose of the code
more clearly and also doesn't need to create a new copy of the list or
fully iterate through a list to complete.

Additionally there are a couple instances of using `any` and `all` instead
of the `List.filter |> List.isEmpty |> not` sequence, or variations of
them. These are being replaced for the same reason.