Minify build in production to reduce bundle size
Exosphere currently serves its main JavaScript bundle in its minimally-optimized
form, passing the output of `elm make --optimize` directly. This means that we've
been sending a 2MB payload to visitors of the site, something that's particularly
rough on the mobile experience (battery usage and download speeds).

In this patch we're introducing `uglifyfs` into the build pipeline as recommended
by the Elm project itself. `uglifyjs` removes whitespace, renames variables, and
performs other optimizations to shrink the size of the built JS bundle.

Specifically this patch brings the output bundle down from 1.8 MB to 407 KB.
In and of itself this helps but is insufficient for the client experience. The
assets are being served without web-server compression. Turning on web server
compression to the browser should result in a final transmitted size drop from
288 KB to 131 KB. _This is left for future work_ however and server-size
compression will not be addressed in this patch.

There should be no functional or visual changes in this patch, only a reduction
in size of the output JS bundle.