Store project description in the model
The only time we get the project description is in the GET call to `/v3/auth/projects`, which we use with an unscoped auth token to ask for a list of project scopes.
(This API call returns the description field for each project, though it is not a documented field in the API docs.)

So, in order to store the description for each scoped project in the model, we need to pass it all the way through the Cmd to get a project-scoped auth token, and the resulting Msg.

For reference, the description field would appear here if it were documented:
https://docs.openstack.org/api-ref/identity/v3/?expanded=get-available-project-scopes-detail#get-available-project-scopes