Rocky instead of CentOS for push button clusters
This relies on the `rocky-linux` branch of the CRI_Jetstream_Cluster
repo. Once/if that is merged to the `main` branch in we should change
the branch in `Helpers.Helpers.createClusterYaml` accordingly.

Other changes introduced:

- Pipe output from `cluster_create_local.sh` to `local_create.log`
- Keep Docker installed on head node by passing `-d` to
`cluster_create_local.sh`. This fixes problem where Guacamole did not
work on the head node.