Fetch volume snapshot data and store in state
When fetching volume data from Cinder, also fetch the volume snapshot
data and store the sizes of those volumes in app state.

This will be used to display the portion of the used storage quota
that those snapshots represent. In contrast, the only explanation
currently for that quota is the individual sizes of the individual
volumes, which won't add up to the total (on account of the missing
snapshot data).