humanBytes: Fix mis-conversion of GiB to bytes and MiB to bytes
When `humanNumber` calls `humanBytes` it converts the native unit of
the data type passed into it into raw byte counts so that `humanBytes`
can perform the appropriate conversion to a displayable format.

Unfortunately when that function was introduced in e1eb0551d4dff398e21762503ea1093200465ca9
it was overlooked that the conversion was performed wrongly, converting
what would be a _gigabyte_ value instead of a _gigibyte_ value.

In this patch we're replacing the conversion factor so that input values
in gigibytes or mebibytes will properly convert through `humanBytes` into
the expected values.

Furthermore we're extending the rounding to at most one decimal digit
when displaying the storage total because previously, with zero digits,
it was rounding values such as 1500 MiB to 2 TB.