Popover: Emphasize in dark mode without drop shadow
Resolves !882

The `Popover` component draws visual distinction from the background content over
which it displays through a drop shadow. Unfortunately, in dark mode views this
shadow blends into the page background and the `Popover` loses its distinction.

In this patch we're replacing the shadow with a highlighted border around the
`Popover` and a mildly-subdued background color within it. The choice for the
background color was based on a quick visual inspection of the color palette
and is not important in itself. Reusing an existing color was done for the
pragmatic purpose of keeping small the scope of this change.

However, also in this change is the inlining of the "active theme" into the
`ExoPalette`. The term "active theme" means the selected light or dark mode
based on whether someone has chosen a specific view mode, or if not, whatever
the OS reports to the browser. While the rest of the palette derives from
this theme and therefore doesn't require we store it in state, the `Popover`
rendering occurs conditionally based on that value and therefore including
it is a practical way to avoid broader refactors.