Image Preload: Replace JS preloading with HTML preloading
This patch replaces some JavaScript-based image preloading code with a
list of `<link rel=preload>` elements. Although the JavaScript was inline
and would evaluate as soon as it's encountered, it's still JavaScript and
still needs to evaluate before preloading the image files.

The HTML version relies on the declarative markup to indicate these resource
needs and theoretically will trigger fetches earlier than would happen via
JS, as the browser need only parse the HTML to know when to start sending
fetches. This is unlikely to be measurable or significant due to the fact
that the JS code was inline at the same location in the document.

Also in this patch is the removal of two `jetstream1-` logo files which
are no longer in the project. These assets were removed on Aug. 1, 2022
in 9569c1c1b7997e0c31e8e6337aab9a13cd1ae006 but these preload directives
were left in. Fetches for the assets currently result in a 404.