App Boot: Send flags as JSON-serialized string for custom decoding
Elm special-cases the intial decoding of `Flags`, which works when those
flags are in control of the developer of the running app. Exosphere is
customizable and so flags are out of control of the code in this repo
and therefore when misconfigured by a cloud operator or Exosphere
application deployer will lead to an app crash in the browser.

The reason is that there's no way to recover from a failure to decode
the special-cased `Flags` type.

In this patch we're changing `Flags` so that it sends a JSON-encoded
version of the app flags at boot which is subsequently decoded through
traditional JSON decoding techniques. This provides the opportunity to
return `Err AppError` and follow the existing flow for failing to boot
the app.

Props: @cmart for finding and fixing a linting issue.