Adds a brief explainer with links to cloud config documentation.
Resolves #915

When presented with the cloud config it was confusing to me what I should write
in the file.  This patch adds a brief description with links to the
documentation for the config file so as to provide a basic pointer to learn more.

Props @julianpistorius for adding a warning alert dialog and other help with styling.