Add relative time element which reveals full time on hover.
In several places in the app we're displaying a relative time, such as
"12 hours ago." Unfortunately this hides important information like
when exactly that was. This problem grows the further back a relative
date is. For example, "a month ago" leaves a large gap in understanding.

In this patch we're creating a new `relativeTimeElement` which will render
a `<time>` HTML element whose inner markup describes the relative time in
a readable fashion (as the app has been doing) but will also show the full
time string on hover (as well as convey it in the `datetime` attribute).