Remove support for "auto" network selection
Previously the CreateServer messages have relied on a stringly-typed version of
the chosen network UUID for selecting which network to attach an instance. This
is because the value comes from an HTML form with string output that previously
supported a special value "auto". Unfortunately that creates ambiguity on the
types because this value was not a UUID but rather a kind of sentinel
indicating that it's something else.

Support for "auto" was effectively removed in !410, but this code was left in place.

In this patch we're removing the code handling "auto" because it shouldn't be
possible within the app to generate that value, and the only supported network
configs are actual UUID values.