Add icon abstraction for FeatherIcons
In a sense this work continues what was done in !538.

While working in random parts of the codebase I found a repeated pattern of
creating the same cumbersome setup for using `FeatherIcon` icons and it
seemed noisy with incidental details.

Because of this this patch creates two new helper functions to standardize
and simplify the mechanism for rendering icons. One need not dive into how
a `FeatherIcon` is transformed into an `Element.Element`, only that the
important bits are _which_ icon is to be used, and optionally, which Elm-UI
attributes to apply to the icon. A second helper function is provided for
the common use-case of needing an icon with no attributes, but rendered
at a specific size.