Remove husky, switching to pre-commit
This is a developer experience change, improving the commit lifecycle.
Hooks are run during the `git commit` phase, and may be skipped with
`git commit [-n/--no-verify]`.

In this commit hooks run should be equivalent to the prior husky hooks.
The major change here is hooks will only be run against the
staged/committing changes, so unfinished/unstaged code will be ignored.