Add elm-review NoDeprecated rule
This rule errors on the use of any deprecated modules or functions
annotated with a `-- @deprecated` comment

We currently use no deprecated modules or functions, so no code changes
are required.

I feel this rule is a good value add, with no real downside right now,
and could easily be removed in the future if we find it problematic

```
-- ELM-REVIEW ERROR ----------------------------------- src/Example.elm:5:8

NoDeprecated: Found new usage of deprecated element

5| some = deprecated
          ^^^^^^^^^^

This element was marked as deprecated and should not be used anymore.

Please check its documentation to know the alternative solutions.
```