Add elm-review rule NoMissingTypeExpose
This rule errors when a private type is used in an exposed declaration.

```
-- ELM-REVIEW ERROR --------------------------------- src/View/Helpers.elm:176:9

(fix) NoMissingTypeExpose: Private type `Edges` should be exposed

176| edges : Edges
             ^^^^^
177| edges =

Users of this module will not be able to annotate a value of this type if they
wanted to. You should expose this type or an alias of this type.
```

This one has less short term benefit, but having an auto-fix is a low cost
add that could lead to long term benefits