Add elm-review rule NoPrematureLetComputation
This rule forbids declarations in `let` bodies that are only sometimes used later

From the rule's documentation:

```
    someFunction n =
        let
            value =
                expensiveComputation n
        in
        if needToCompute then
            value + 1

        else
            0

    someFunction n =
        if needToCompute then
            let
                value =
                    expensiveComputation n
            in
            value + 1

        else
            0
```

This would be difficult, except it includes an auto-fix
We have 96 "errors" as is

```
-- ELM-REVIEW ERROR ------------------------------------ src/View/View.elm:141:9

(fix) NoPrematureLetComputation: Let value was declared prematurely

141|         headerContainerAttrs =
             ^^^^^^^^^^^^^^^^^^^^
142|             [ Background.color <|

This value is only used in some code paths, and it can therefore be computed
unnecessarily.

Try moving it closer to where it is needed, I recommend to move it to line 172.
```