Add elm-review rule NoSimpleLetBody
This rule forbids simple `let .. in` blocks where the expression body is
simply a reference to a declaration in the `let`

For example,

```
add a b =
    let
        value =
            a + b
    in
        value
```

We currently show 17 errors for this rule, which can be auto-fixed

```
-- ELM-REVIEW ERROR -------------------------------- src/State/State.elm:2676:21

(fix) NoSimpleLetBody: The referenced value should be inlined.

2675|                     in
2676|                     newSharedModel
                          ^^^^^^^^^^^^^^
2677|             in

The name of the value is redundant with the surrounding expression.

If you believe that the expression needs a name because it is too complex,
consider splitting the expression up more or extracting it to a new function.
```