Add elm-review rule NoUnnecessaryTrailingUnderscore
This rule forbids variable names ending with an underscore when there is
 no "need" for said underscore. This assumes the underscore was used to
 avoid a variable shadowing issue, or possibly to avoid a "keyword"

The intended "valid" usage would be

```
viewName name =
      case name of
        Just name_ -> ...
        Nothing -> ...
```

When variables are renamed or code is refactored and these variables are
no longer shadowed, there would be no more need for the underscore.

We currently have 71 places where this rule would make changes.

I found two possible cases where this rule can trigger: top level
declarations which should use "better" variable names instead of
underscores, and variables in blocks that no longer need shadowing.
Both errors follow

```
-- ELM-REVIEW ERROR ------------------------------------- src/View/View.elm:97:1

NoUnnecessaryTrailingUnderscore: Top-level declaration names should not end with
an underscore

96| view_ : OuterModel -> Html.Html OuterMsg
97| view_ outerModel =
    ^^^^^
98|     let

A trailing underscore "_" is often used to prevent shadowing issues, but
top-level declarations should not resolve these issues in that manner.

─────────────────────────────────────────────────────── src/View/View.elm:132:30

(fix) NoUnnecessaryTrailingUnderscore: project_ should not end with an
underscore

132|                         Just project_ ->
                                  ^^^^^^^^
133|                             ( Just <| projectHeaderView context project_

It looks like this was used to avoid a shadowing issue, but the variable it
would have clashed with is not present in the scope of where this variable was
declared anymore. You should rename the variable and remove the underscore.

Note that this may not be a safe change, in that renaming may clash with a value
declared deeper in the expression, but I think it's less confusing to have the
nested variable have a trailing underscore rather than the variable declared
higher up.
```