Simplify: immediately evaluated anonymous function
This one is a special case, instead of the suggested fixes I was able to
just update the definition of projectSetServerLoading to swap the
ServerUuid and Project arguments, allowing easy pipelining and keeping
with the common code style of Exosphere.

The other easy fix in this case could have been

```
    GetterSetters.projectSetServerLoading project serverUuid
        |> GetterSetters.modelUpdateProject model
```

```
-- ELM-REVIEW ERROR ------------------------- src/Rest/ApiModelHelpers.elm:65:21

Simplify: Anonymous function is immediately invoked

64|             ( project
65|                 |> (\p -> GetterSetters.projectSetServerLoading p serverUuid)
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
66|                 |> GetterSetters.modelUpdateProject model

This expression defines a function which then gets called directly afterwards,
which overly complexifies the intended computation.

While there are reasonable uses for this in languages like JavaScript, the same
benefits aren't there in Elm because of not allowing name shadowing.

Here are a few ways you can simplify this:

- Remove the lambda and reference the arguments directly instead of giving them
new names
- Remove the lambda and use let variables to give names to the current arguments
- Extract the lambda to a named function (at the top-level or defined in a let
expression)

──────────────────────────────────────────── src/Rest/ApiModelHelpers.elm:104:21

Simplify: Anonymous function is immediately invoked

103|             ( project
104|                 |> (\p -> GetterSetters.projectSetServerEventsLoading p serverUuid)
                         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
105|                 |> GetterSetters.modelUpdateProject model

This expression defines a function which then gets called directly afterwards,
which overly complexifies the intended computation.

While there are reasonable uses for this in languages like JavaScript, the same
benefits aren't there in Elm because of not allowing name shadowing.

Here are a few ways you can simplify this:

- Remove the lambda and reference the arguments directly instead of giving them
new names
- Remove the lambda and use let variables to give names to the current arguments
- Extract the lambda to a named function (at the top-level or defined in a let
expression)
```