Add elm-review rule NoUnused.CustomTypeConstructors
This detects custom types where some constructors are never created

```
-- ELM-REVIEW ERROR ---------------------- src/Style/Widgets/IconButton.elm:16:7

(fix) NoUnused.CustomTypeConstructors: Type constructor `After` is not used.

15|     = Before
16|     | After
          ^^^^^

This type constructor is never used. It might be handled everywhere it might
appear, but there is no location where this value actually gets created.
```

In this case, Flow.After is never created, although it is handled
This rule might be a little troublesome for us, but is handled in a case
statement at src/Style/Widgets/IconButton.elm:52:21

```
case iconPlacement of
    Before ->
        [ iconUI, labelUI ]

    After ->
        [ labelUI, iconUI ]
```

This one could be useful later, and I don't feel it should be removed

There are several other places, such as several SharedMsg types

```
-- ELM-REVIEW ERROR ------------------------------- src/Types/SharedMsg.elm:59:7

NoUnused.CustomTypeConstructors: Type constructor `RequestServers` is not used.

58|     | ServerMsg OSTypes.ServerUuid ServerSpecificMsgConstructor
59|     | RequestServers
          ^^^^^^^^^^^^^^
60|     | RequestCreateServer HelperTypes.CreateServerPageModel OSTypes.NetworkUuid OSTypes.FlavorId

This type constructor is never used. It might be handled everywhere it might
appear, but there is no location where this value actually gets created.

─────────────────────────────────────────────────── src/Types/SharedMsg.elm:66:7

NoUnused.CustomTypeConstructors: Type constructor `RequestKeypairs` is not used.

65|     | RequestDetachVolume OSTypes.VolumeUuid
66|     | RequestKeypairs
          ^^^^^^^^^^^^^^^
67|     | RequestCreateKeypair OSTypes.KeypairName OSTypes.PublicKey

This type constructor is never used. It might be handled everywhere it might
appear, but there is no location where this value actually gets created.

────────────────────────────────────────────────── src/Types/SharedMsg.elm:116:7

NoUnused.CustomTypeConstructors: Type constructor `RequestServer` is not used.

115| type ServerSpecificMsgConstructor
116|     = RequestServer
           ^^^^^^^^^^^^^
117|     | RequestDeleteServer Bool

This type constructor is never used. It might be handled everywhere it might
appear, but there is no location where this value actually gets created.
```